moved the class to its own file
